requirements.txt
# Core dependencies for Sugarcane Disease Classification

# Ultralytics YOLO (includes PyTorch, OpenCV, etc.)
ultralytics>=8.1.0

# Kaggle dataset download
kagglehub>=0.1.4

# Data manipulation
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model export formats
onnx>=1.15.0
onnxruntime>=1.16.0

# TFLite conversion (optional, for local export)
tensorflow>=2.15.0

# Progress bars
tqdm>=4.65.0

# Additional utilities
pyyaml>=6.0
scikit-learn>=1.3.0
